Create component for matrix
